Move AddLock to fixed position in MenuWrapper
