+++ src/styles/components/_menuWrapper.scss
@@ -1,16 +1,6 @@
 .menuWrapper__container {
 }
 
-.menuWrapper__container--addAuth {
-    justify-content: space-between;
-    display: flex
-}
-
-.menuWrapper__button--addLock {
-    background: transparent;
-    border: none;
-}
-
 .menuWrapper__container--menuItems {
     display: flexbox;
     flex-wrap: wrap;
